feat: Update password with reset token
The code changes add a new method  to the  class in the  file. This method allows users to update their password by providing a reset token and a new password. It retrieves the user associated with the reset token, hashes the new password, and updates the user's hashed password and reset token.

This commit message follows the established convention of using a prefix to indicate the type of change. In this case,  is used to indicate a new feature addition.